float, meausrements ILayoutNode, OptAlgoGraphRenderer, Default attributes random city, move final layout to centerPosiition (0,0),

+++ Assets/SEECity/DataModel/CoseGraphSettings.cs
@@ -10,12 +10,12 @@ namespace SEE.DataModel
     {
         public int EdgeLength = CoseLayoutSettings.Edge_Length;
         public bool UseSmartIdealEdgeCalculation = CoseLayoutSettings.Use_Smart_Ideal_Edge_Calculation;
-        public double PerLevelIdealEdgeLengthFactor = CoseLayoutSettings.Per_Level_Ideal_Edge_Length_Factor;
+        public float PerLevelIdealEdgeLengthFactor = CoseLayoutSettings.Per_Level_Ideal_Edge_Length_Factor;
         public bool Incremental = CoseLayoutSettings.Incremental;
         public bool UseSmartRepulsionRangeCalculation = CoseLayoutSettings.Use_Smart_Repulsion_Range_Calculation;
-        public double GravityStrength = CoseLayoutSettings.Gravity_Strength;
-        public double CompoundGravityStrength = CoseLayoutSettings.Compound_Gravity_Strength;
-        public double RepulsionStrength = CoseLayoutSettings.Repulsion_Strength;
+        public float GravityStrength = CoseLayoutSettings.Gravity_Strength;
+        public float CompoundGravityStrength = CoseLayoutSettings.Compound_Gravity_Strength;
+        public float RepulsionStrength = CoseLayoutSettings.Repulsion_Strength;
         public bool multiLevelScaling = CoseLayoutSettings.Multilevel_Scaling;
         public Dictionary<string, bool> ListDirToggle = new Dictionary<string, bool>();
         public int index = 0;
